Fix atomic value calculation for stake/transfer
The logic underlying the conversion of input value into atomic value was
wrong: for some values, `parseFloat(amount).toFixed(9) * 1e9` doesn't
result in an integer because of numeric imprecision, and so a
non-integer value gets passed into the RPC wallet which chokes on it.

For example:

    ((parseFloat("17199.899334361")).toFixed(9)*1e9).toString()

gives

    17199899334360.998

This fixes the code to multiply by 1e9 *before* rounding to an atomic
unit so that we always get an integer value.